refactor: Define internal subpackages
These subpackages are specific to `opinions` and they are not useful for the
external users.
