+++ reddit_test.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/macie/opinions/ensure"
-	"github.com/macie/opinions/http"
+	"github.com/macie/opinions/internal/ensure"
+	"github.com/macie/opinions/internal/http"
 )
 
 func ExampleSearchReddit() {
